Organize queries by type
