--- db/queries/getOrderById.js
@@ -1,13 +1,0 @@
-const db = require('../connection');
-
-const getOrderById = (Id) => {
-  return db.query(`
-  SELECT * FROM orders
-  WHERE id = $1;
-  `, [Id])
-    .then(data => {
-      return data.rows[0];
-    });
-};
-
-exports.getOrderById = getOrderById;
