+++ db/queries/users.js
@@ -8,3 +8,14 @@ const getUsers = () => {
 };
 
 exports.getUsers =  getUsers;
+
+const getUserById = (user) => {
+  return db.query(`
+  SELECT * FROM users
+  WHERE id = $1;`, [user.id])
+    .then(data => {
+      return data.rows[0];
+    });
+};
+
+exports.getUserById =  getUserById;
